+++ DataMaps/ModEntry.cs
@@ -4,6 +4,7 @@ using Pathoschild.Stardew.Common;
 using Pathoschild.Stardew.DataMaps.DataMaps;
 using Pathoschild.Stardew.DataMaps.Framework;
 using Pathoschild.Stardew.DataMaps.Framework.Integrations.BetterSprinklers;
+using Pathoschild.Stardew.DataMaps.Framework.Integrations.Cobalt;
 using Pathoschild.Stardew.DataMaps.Framework.Integrations.PelicanFiber;
 using StardewModdingAPI;
 using StardewModdingAPI.Events;
@@ -27,11 +28,14 @@ namespace Pathoschild.Stardew.DataMaps
         /// <summary>The available data maps.</summary>
         private IDataMap[] Maps;
 
-        /// <summary>Handles the logic for integrating with the Pelican Fiber mod.</summary>
+        /// <summary>Handles access to the Pelican Fiber mod.</summary>
         private PelicanFiberIntegration PelicanFiber;
 
-        /// <summary>Handles the logic for integrating with the Better Sprinklers mod.</summary>
+        /// <summary>Handles access to the Better Sprinklers mod.</summary>
         private BetterSprinklersIntegration BetterSprinklers;
+
+        /// <summary>Handles access to the Cobalt mod.</summary>
+        private CobaltIntegration Cobalt;
 
 
         /*********
@@ -41,20 +45,12 @@ namespace Pathoschild.Stardew.DataMaps
         /// <param name="helper">Provides methods for interacting with the mod directory, such as read/writing a config file or custom JSON files.</param>
         public override void Entry(IModHelper helper)
         {
-            // initialise
+            // read config
             this.Config = helper.ReadConfig<ModConfig>();
-            this.PelicanFiber = new PelicanFiberIntegration(helper.ModRegistry, helper.Reflection, this.Monitor);
-            this.BetterSprinklers = new BetterSprinklersIntegration(helper.ModRegistry, helper.Reflection, this.Monitor);
-            this.Maps = new IDataMap[]
-            {
-                new AccessibilityMap(helper.Translation),
-                new ScarecrowMap(helper.Translation),
-                new SprinklerMap(helper.Translation, this.BetterSprinklers),
-                new JunimoHutMap(helper.Translation, this.PelicanFiber)
-            };
 
             // hook up events
             SaveEvents.AfterReturnToTitle += this.SaveEvents_AfterReturnToTitle;
+            GameEvents.FirstUpdateTick += this.GameEvents_FirstUpdateTick;
             GameEvents.SecondUpdateTick += this.GameEvents_SecondUpdateTick;
             InputEvents.ButtonPressed += this.InputEvents_ButtonPressed;
         }
@@ -63,6 +59,25 @@ namespace Pathoschild.Stardew.DataMaps
         /*********
         ** Private methods
         *********/
+        /// <summary>The method invoked on the first game update tick.</summary>
+        /// <param name="sender">The event sender.</param>
+        /// <param name="e">The event arguments.</param>
+        private void GameEvents_FirstUpdateTick(object sender, EventArgs e)
+        {
+            IModHelper helper = this.Helper;
+
+            this.PelicanFiber = new PelicanFiberIntegration(helper.ModRegistry, helper.Reflection, this.Monitor);
+            this.BetterSprinklers = new BetterSprinklersIntegration(helper.ModRegistry, helper.Reflection, this.Monitor);
+            this.Cobalt = new CobaltIntegration(helper.ModRegistry, this.Monitor);
+            this.Maps = new IDataMap[]
+            {
+                new AccessibilityMap(helper.Translation),
+                new ScarecrowMap(helper.Translation),
+                new SprinklerMap(helper.Translation, this.BetterSprinklers, this.Cobalt),
+                new JunimoHutMap(helper.Translation, this.PelicanFiber)
+            };
+        }
+
         /// <summary>The method invoked when the player returns to the title screen.</summary>
         /// <param name="sender">The event sender.</param>
         /// <param name="e">The event arguments.</param>