+++ DataMaps/release-notes.md
@@ -1,5 +1,12 @@
 # Release notes
+### 1.1
+* Added support for...
+  * Cobalt's sprinklers.
+* Fixed deprecated API usage.
+
 ### 1.0
 * Initial version.
 * Added Junimo huts, scarecrows, sprinklers, and accessibility data maps.
-* Added compatibility with Better Sprinklers and Pelican Fiber.
+* Added support for...
+  * Better Sprinklers' custom sprinklers;
+  * Pelican Fiber's custom build menu.